requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
filelock==3.25.2
fsspec==2026.3.0
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.4.2
ninja==1.13.0
nncf==3.0.0
numpy==2.2.6
onnx==1.21.0
openvino==2026.0.0
openvino-telemetry==2025.2.0
openvino-tokenizers==2026.0.0.0
optimum==2.1.0
optimum-intel==1.27.0
optimum-onnx==0.1.0
packaging==26.0
pandas==2.3.3
pillow==12.1.1
protobuf==7.34.1
psutil==7.2.2
pydot==3.0.4
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.3.32
requests==2.33.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sympy==1.14.0
tabulate==0.10.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==4.57.6
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
